refactor: GesturableGraphView로 분리
