+++ Sources/GesturableGraph/GesturableGraph.swift
@@ -1,8 +1,23 @@
+//
+//  GesturableGraph.swift
+//
+//
+//  Created by 박재우 on 10/11/23.
+//
+
 import UIKit
 
-public final class GesturableGraph: UIView, Gesturable {
-    private(set) public var elements = [Double]()
+public final class GesturableGraph: UIView {
+    private var graph: Graph
 
+    public let gesturableGraphView: GesturableGraphView
+    public let axisYView: AxisYView
+
+    public lazy var elements = graph.elements {
+        didSet {
+            graph.elements = elements
+        }
+    }
     public lazy var distribution = graph.distribution {
         didSet {
             graph.distribution = distribution
@@ -18,32 +33,6 @@ public final class GesturableGraph: UIView, Gesturable {
             graph.padding = padding
         }
     }
-    private lazy var points = {
-        graph.points.map { x, y in
-            CGPoint(x: bounds.width * x, y: bounds.height * y)
-        }
-    }()
-
-    public var line = GraphLine()
-    public var point = GraphPoint()
-    public var area = GraphArea()
-    public var enablePoint = GraphEnablePoint() {
-        didSet {
-            gestureEnableView.updatePointView(width: enablePoint.width,
-                                              color: enablePoint.color)
-        }
-    }
-    public var enableLine = GraphEnableLine() {
-        didSet {
-            gestureEnableView.updateLineView(width: enableLine.width,
-                                              color: enableLine.color)
-        }
-    }
-
-    internal var gesture: CGPoint?
-
-    private var gestureEnableView = GestureEnableView()
-    private var graph: Graph
 
     public init?(elements: [Double]) {
         guard let graph = Graph(elements: elements) else {
@@ -51,18 +40,10 @@ public final class GesturableGraph: UIView, Gesturable {
         }
 
         self.graph = graph
+        self.gesturableGraphView = GesturableGraphView(graph: graph)
+        self.axisYView = AxisYView(top: graph.calibrationTop, bottom: graph.calibrationBottom)
 
         super.init(frame: .zero)
-
-        setUI()
-    }
-
-    override init(frame: CGRect) {
-        self.graph = Graph()
-
-        super.init(frame: frame)
-
-        setUI()
     }
 
     required init?(coder: NSCoder) {
@@ -70,65 +51,22 @@ public final class GesturableGraph: UIView, Gesturable {
     }
 
     private func setUI() {
-        let enableViewidth = max(enableLine.width, enablePoint.width)
-        addSubview(gestureEnableView)
-        gestureEnableView.translatesAutoresizingMaskIntoConstraints = false
-
-        NSLayoutConstraint.activate([
-            gestureEnableView.topAnchor.constraint(equalTo: topAnchor),
-            gestureEnableView.bottomAnchor.constraint(equalTo: bottomAnchor),
-            gestureEnableView.widthAnchor.constraint(equalToConstant: enableViewidth),
-            gestureEnableView.leadingAnchor.constraint(equalTo: leadingAnchor)
-        ])
-
-        let axisYView = AxisYView(top: graph.elements.calibrationTop(ofValue: graph.padding.top)!,
-                                  bottom: graph.elements.calibrationBottom(ofValue: graph.padding.bottom)!)
         addSubview(axisYView)
+        addSubview(gesturableGraphView)
+
         axisYView.translatesAutoresizingMaskIntoConstraints = false
+        gesturableGraphView.translatesAutoresizingMaskIntoConstraints = false
 
         NSLayoutConstraint.activate([
             axisYView.topAnchor.constraint(equalTo: topAnchor),
             axisYView.trailingAnchor.constraint(equalTo: trailingAnchor),
             axisYView.bottomAnchor.constraint(equalTo: bottomAnchor),
+
+            gesturableGraphView.topAnchor.constraint(equalTo: axisYView.arrangedSubviews.first!.centerYAnchor),
+            gesturableGraphView.leadingAnchor.constraint(equalTo: leadingAnchor),
+            gesturableGraphView.trailingAnchor.constraint(equalTo: trailingAnchor),
+            gesturableGraphView.bottomAnchor.constraint(equalTo: axisYView.arrangedSubviews.last!.centerYAnchor)
         ])
     }
 
-    public override func draw(_ rect: CGRect) {
-        super.draw(rect)
-
-        guard let mainContext = UIGraphicsGetCurrentContext() else {
-            return
-        }
-
-        mainContext.saveGState()
-
-        guard let graphPath = graphPath(through: points) else {
-            return
-        }
-
-        fillGraphArea(graphPath, using: points)
-        draw(graphPath)
-        draw(points)
-    }
-
-    public override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
-        gesture = touches.first?.location(in: self)
-
-        if let point = calculatedPoint(in: graph, withSize: bounds) {
-            gestureEnableView.moveTo(point)
-            gestureEnableView.isHidden = false
-        }
-    }
-
-    public override func touchesEnded(_ touches: Set<UITouch>, with event: UIEvent?) {
-        gestureEnableView.isHidden = true
-    }
-
-    public override func touchesMoved(_ touches: Set<UITouch>, with event: UIEvent?) {
-        gesture = touches.first?.location(in: self)
-
-        if let point = calculatedPoint(in: graph, withSize: bounds) {
-            gestureEnableView.moveTo(point)
-        }
-    }
 }
